WIIS-1610: crée les mouvements de stock à validation réception + ajoute "issu de" dans liste mouvements + export csv + centralise initAjaxAutocomplete + fix génération CSV mouvements de stock + fix bug initialisation affichage filtre sup mouvements stock (type) + répare lien entités mouvement de stock/collecte + ajoute lien mouvement stock/collecte à la création du mouvement

+++ src/Controller/ReceptionController.php
@@ -43,6 +43,7 @@ use App\Repository\ReceptionReferenceArticleRepository;
 use App\Repository\TransporteurRepository;
 
 use App\Service\DemandeLivraisonService;
+use App\Service\MouvementStockService;
 use App\Service\ReceptionService;
 use App\Service\AttachmentService;
 use App\Service\ArticleDataService;
@@ -51,7 +52,6 @@ use App\Service\UserService;
 use DateTime;
 use DateTimeZone;
 use Doctrine\ORM\NonUniqueResultException;
-use Exception;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -180,6 +180,11 @@ class ReceptionController extends AbstractController
      */
     private $paramGlobalRepository;
 
+	/**
+	 * @var MouvementStockService
+	 */
+    private $mouvementStockService;
+
     public function __construct(
         ArticleDataService $articleDataService,
         DimensionsEtiquettesRepository $dimensionsEtiquettesRepository,
@@ -203,7 +208,8 @@ class ReceptionController extends AbstractController
         AttachmentService $attachmentService,
         FieldsParamRepository $fieldsParamRepository,
         TransporteurRepository $transporteurRepository,
-        ParametrageGlobalRepository $parametrageGlobalRepository
+        ParametrageGlobalRepository $parametrageGlobalRepository,
+		MouvementStockService $mouvementStockService
     )
     {
         $this->paramGlobalRepository = $parametrageGlobalRepository;
@@ -229,6 +235,7 @@ class ReceptionController extends AbstractController
         $this->articleDataService = $articleDataService;
         $this->transporteurRepository = $transporteurRepository;
         $this->fieldsParamRepository = $fieldsParamRepository;
+        $this->mouvementStockService = $mouvementStockService;
     }
 
 
@@ -619,6 +626,9 @@ class ReceptionController extends AbstractController
             }
 
             $ligneArticle = $this->receptionReferenceArticleRepository->find($data['ligneArticle']);
+
+            if (!$ligneArticle) return new JsonResponse(false);
+
             $reception = $ligneArticle->getReception();
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($ligneArticle);
@@ -1288,7 +1298,6 @@ class ReceptionController extends AbstractController
         }
 
         if ($request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $em = $this->getDoctrine()->getManager();
             $reception = $this->receptionRepository->find($data['id']);
             $listReceptionReferenceArticle = $this->receptionReferenceArticleRepository->findByReception($reception);
 
@@ -1296,14 +1305,14 @@ class ReceptionController extends AbstractController
                 return new JsonResponse('Vous ne pouvez pas finir une réception sans article.');
             } else {
                 if ($data['confirmed'] === true) {
-                    $this->validateReception($em, $reception, $listReceptionReferenceArticle);
+                    $this->validateReception($reception, $listReceptionReferenceArticle);
                     return new JsonResponse(1);
                 } else {
                     $partielle = false;
                     foreach ($listReceptionReferenceArticle as $receptionRA) {
                         if ($receptionRA->getQuantite() !== $receptionRA->getQuantiteAR()) $partielle = true;
                     }
-                    if (!$partielle) $this->validateReception($em, $reception, $listReceptionReferenceArticle);
+                    if (!$partielle) $this->validateReception($reception, $listReceptionReferenceArticle);
                     return new JsonResponse($partielle ? 0 : 1);
                 }
             }
@@ -1311,29 +1320,54 @@ class ReceptionController extends AbstractController
         throw new NotFoundHttpException("404");
     }
 
-    /**
-     * @param $em
-     * @param $reception
-     * @param $listReceptionReferenceArticle
-     * @throws NonUniqueResultException
-     * @throws Exception
-     */
-    public function validateReception($em, $reception, $listReceptionReferenceArticle)
+	/**
+	 * @param Reception $reception
+	 * @param ReceptionReferenceArticle[] $listReceptionReferenceArticle
+	 * @throws NonUniqueResultException
+	 */
+    private function validateReception($reception, $listReceptionReferenceArticle)
     {
+    	$em = $this->getDoctrine()->getManager();
         $statut = $this->statutRepository->findOneByCategorieNameAndStatutName(Reception::CATEGORIE, Reception::STATUT_RECEPTION_TOTALE);
-        foreach ($listReceptionReferenceArticle as $receptionRA) {
+		$now = new DateTime('now', new DateTimeZone('Europe/Paris'));
+
+		foreach ($listReceptionReferenceArticle as $receptionRA) {
             $referenceArticle = $receptionRA->getReferenceArticle();
             if ($referenceArticle->getTypeQuantite() === ReferenceArticle::TYPE_QUANTITE_REFERENCE) {
                 $referenceArticle->setQuantiteStock($referenceArticle->getQuantiteStock() + $receptionRA->getQuantite());
-            }
-        }
 
-        $now = new DateTime('now', new DateTimeZone('Europe/Paris'));
+                $mouvementStock = new MouvementStock();
+                $mouvementStock
+					->setUser($this->getUser())
+					->setEmplacementTo($referenceArticle->getEmplacement())
+					->setQuantity($receptionRA->getQuantite())
+					->setRefArticle($referenceArticle)
+					->setType(MouvementStock::TYPE_ENTREE)
+					->setReceptionOrder($reception)
+					->setDate($now);
+				$em->persist($mouvementStock);
+            } else {
+				$articles = $receptionRA->getArticles();
+				foreach ($articles as $article) {
+					$mouvementStock = new MouvementStock();
+					$mouvementStock
+						->setUser($this->getUser())
+						->setEmplacementTo($article->getEmplacement())
+						->setQuantity($article->getQuantite())
+						->setArticle($article)
+						->setType(MouvementStock::TYPE_ENTREE)
+						->setReceptionOrder($reception)
+						->setDate($now);
+					$em->persist($mouvementStock);
+				}
+			}
+        }
 
         $reception
             ->setStatut($statut)
             ->setDateFinReception($now)
             ->setDateCommande($now);
+
         $em->flush();
     }
 
@@ -1530,116 +1564,117 @@ class ReceptionController extends AbstractController
         throw new NotFoundHttpException("404");
     }
 
-    /**
-     * @Route("/valider_lot", name="validate_lot", options={"expose"=true}, methods={"GET", "POST"})
-     * @throws Exception
-     */
-    public function validateLot(Request $request)
-    {
-        if ($request->isXmlHttpRequest() && $dataContent = json_decode($request->getContent(), true)) {
-            $em = $this->getDoctrine()->getManager();
-            $response = [];
-            $response['refs'] = [];
-            $response['barcodesLabel'] = [];
-            $dimension = $this->dimensionsEtiquettesRepository->findOneDimension();
-            if ($dimension && !empty($dimension->getHeight()) && !empty($dimension->getWidth())) {
-                $response['height'] = $dimension->getHeight();
-                $response['width'] = $dimension->getWidth();
-                $response['exists'] = true;
-            } else {
-                $response['exists'] = false;
-            }
-            $qtt = 0;
-            for ($i = 0; $i < count($dataContent['quantiteLot']); $i++) {
-                for ($j = 0; $j < $dataContent['quantiteLot'][$i]; $j++) {
-                    $qtt += $dataContent['tailleLot'][$i];
-                }
-            }
-            $ligne = $this->receptionReferenceArticleRepository->find(intval($dataContent['ligne']));
-            if ($qtt + $ligne->getQuantite() > $ligne->getQuantiteAR()) {
-                $response['exists'] = false;
-            } else {
-                $ligne->setQuantite($ligne->getQuantite() + $qtt);
-            }
-
-            if ($response['exists'] === true) {
-                $refArticle = $this->referenceArticleRepository->findOneByReference($dataContent['refArticle']);
-
-                $date = new DateTime('now', new DateTimeZone('Europe/Paris'));
-                $formattedDate = $date->format('ym');
-                $references = $this->articleRepository->getReferencesByRefAndDate($refArticle, $formattedDate);
-
-                $highestCpt = 0;
-                foreach ($references as $reference) {
-                    $cpt = (int)substr($reference, -5, 5);
-                    if ($cpt > $highestCpt) $highestCpt = $cpt;
-                }
-                $counter = $highestCpt + 1;
-                $wantBL = $this->paramGlobalRepository->findOneByLabel(ParametrageGlobal::INCLUDE_BL_IN_LABEL);
-                for ($i = 0; $i < count($dataContent['quantiteLot']); $i++) {
-                    for ($j = 0; $j < $dataContent['quantiteLot'][$i]; $j++) {
-
-                        $toInsert = new Article();
-                        $statut = $this->statutRepository->findOneByCategorieNameAndStatutName(Article::CATEGORIE, Article::STATUT_ACTIF);
-                        $ligne = $this->receptionReferenceArticleRepository->find(intval($dataContent['ligne']));
-                        $reception = $this->receptionRepository->find($dataContent['receptionId']);
-                        $articleFournisseur = new ArticleFournisseur();
-                        $articleFournisseur
-                            ->setReferenceArticle($refArticle)
-                            ->setFournisseur($reception->getFournisseur())
-                            ->setReference($refArticle->getReference())
-                            ->setLabel($ligne->getCommande());
-                        $em->persist($articleFournisseur);
-
-                        $formattedCounter = sprintf('%05u', $counter);
-                        $toInsert
-                            ->setLabel($ligne->getCommande())
-                            ->setConform(true)
-                            ->setStatut($statut)
-                            ->setReference($refArticle->getReference() . $formattedDate . $formattedCounter)
-                            ->setQuantite(max(intval($dataContent['tailleLot'][$i]), 0))// protection contre quantités négatives
-                            ->setArticleFournisseur($articleFournisseur)
-                            ->setReceptionReferenceArticle($ligne)
-                            ->setType($refArticle->getType())
-                            ->setBarCode($this->articleDataService->generateBarCode());
-
-                        $newMouvement = new MouvementStock();
-                        $newMouvement
-                            ->setUser($this->getUser())
-                            ->setArticle($toInsert)
-                            ->setDate(new DateTime('now', new DateTimeZone('Europe/Paris')))
-                            ->setType(MouvementStock::TYPE_ENTREE)
-                            ->setQuantity($toInsert->getQuantite());
-                        $em->persist($newMouvement);
-
-                        $em->persist($toInsert);
-                        $em->flush();
-                        $articles = $this->articleRepository->getRefAndLabelRefAndArtAndBarcodeAndBLById($toInsert->getId());
-                        $wantedIndex = 0;
-                        foreach($articles as $key => $articleWithCL) {
-                            if ($articleWithCL['cl'] === ChampLibre::SPECIC_COLLINS_BL) {
-                                $wantedIndex = $key;
-                                break;
-                            }
-                        }
-                        $articleArray = $articles[$wantedIndex];
-                        array_push($response['refs'], $toInsert->getBarCode());
-                        array_push($response['barcodesLabel'], $this->renderView('article/barcodeLabel.html.twig', [
-                            'refRef' => $toInsert->getArticleFournisseur()->getReferenceArticle()->getReference(),
-                            'refLabel' => $toInsert->getArticleFournisseur()->getReferenceArticle()->getLibelle(),
-                            'artLabel' => $toInsert->getLabel(),
-                            'artBL' => $wantBL ? $wantBL->getParametre() && $articleArray['cl'] === ChampLibre::SPECIC_COLLINS_BL ? $articleArray['bl'] : null : null,
-                        ])
-                        );
-                        $counter++;
-                    }
-                }
-            }
-            $em->flush();
-            return new JsonResponse($response);
-        }
-        throw new NotFoundHttpException("404");
-    }
+//    /**
+//     * @Route("/valider_lot", name="validate_lot", options={"expose"=true}, methods={"GET", "POST"})
+//     * @throws Exception
+//     */
+//    public function validateLot(Request $request)
+//    {
+//        if ($request->isXmlHttpRequest() && $dataContent = json_decode($request->getContent(), true)) {
+//            $em = $this->getDoctrine()->getManager();
+//            $response = [];
+//            $response['refs'] = [];
+//            $response['barcodesLabel'] = [];
+//            $dimension = $this->dimensionsEtiquettesRepository->findOneDimension();
+//            if ($dimension && !empty($dimension->getHeight()) && !empty($dimension->getWidth())) {
+//                $response['height'] = $dimension->getHeight();
+//                $response['width'] = $dimension->getWidth();
+//                $response['exists'] = true;
+//            } else {
+//                $response['exists'] = false;
+//            }
+//            $qtt = 0;
+//            for ($i = 0; $i < count($dataContent['quantiteLot']); $i++) {
+//                for ($j = 0; $j < $dataContent['quantiteLot'][$i]; $j++) {
+//                    $qtt += $dataContent['tailleLot'][$i];
+//                }
+//            }
+//            $ligne = $this->receptionReferenceArticleRepository->find(intval($dataContent['ligne']));
+//            if ($qtt + $ligne->getQuantite() > $ligne->getQuantiteAR()) {
+//                $response['exists'] = false;
+//            } else {
+//                $ligne->setQuantite($ligne->getQuantite() + $qtt);
+//            }
+//
+//            if ($response['exists'] === true) {
+//                $refArticle = $this->referenceArticleRepository->findOneByReference($dataContent['refArticle']);
+//
+//                $date = new DateTime('now', new DateTimeZone('Europe/Paris'));
+//                $formattedDate = $date->format('ym');
+//                $references = $this->articleRepository->getReferencesByRefAndDate($refArticle, $formattedDate);
+//
+//                $highestCpt = 0;
+//                foreach ($references as $reference) {
+//                    $cpt = (int)substr($reference, -5, 5);
+//                    if ($cpt > $highestCpt) $highestCpt = $cpt;
+//                }
+//                $counter = $highestCpt + 1;
+//                $wantBL = $this->paramGlobalRepository->findOneByLabel(ParametrageGlobal::INCLUDE_BL_IN_LABEL);
+//                for ($i = 0; $i < count($dataContent['quantiteLot']); $i++) {
+//                    for ($j = 0; $j < $dataContent['quantiteLot'][$i]; $j++) {
+//
+//                        $toInsert = new Article();
+//                        $statut = $this->statutRepository->findOneByCategorieNameAndStatutName(Article::CATEGORIE, Article::STATUT_ACTIF);
+//                        $ligne = $this->receptionReferenceArticleRepository->find(intval($dataContent['ligne']));
+//                        $reception = $this->receptionRepository->find($dataContent['receptionId']);
+//                        $articleFournisseur = new ArticleFournisseur();
+//                        $articleFournisseur
+//                            ->setReferenceArticle($refArticle)
+//                            ->setFournisseur($reception->getFournisseur())
+//                            ->setReference($refArticle->getReference())
+//                            ->setLabel($ligne->getCommande());
+//                        $em->persist($articleFournisseur);
+//
+//                        $formattedCounter = sprintf('%05u', $counter);
+//                        $toInsert
+//                            ->setLabel($ligne->getCommande())
+//                            ->setConform(true)
+//                            ->setStatut($statut)
+//                            ->setReference($refArticle->getReference() . $formattedDate . $formattedCounter)
+//                            ->setQuantite(max(intval($dataContent['tailleLot'][$i]), 0))// protection contre quantités négatives
+//                            ->setArticleFournisseur($articleFournisseur)
+//                            ->setReceptionReferenceArticle($ligne)
+//                            ->setType($refArticle->getType())
+//                            ->setBarCode($this->articleDataService->generateBarCode());
+//
+//                        $newMouvement = new MouvementStock();
+//                        $newMouvement
+//                            ->setUser($this->getUser())
+//                            ->setArticle($toInsert)
+//                            ->setDate(new DateTime('now', new DateTimeZone('Europe/Paris')))
+//                            ->setType(MouvementStock::TYPE_ENTREE)
+//							->setReceptionOrder()
+//                            ->setQuantity($toInsert->getQuantite());
+//                        $em->persist($newMouvement);
+//
+//                        $em->persist($toInsert);
+//                        $em->flush();
+//                        $articles = $this->articleRepository->getRefAndLabelRefAndArtAndBarcodeAndBLById($toInsert->getId());
+//                        $wantedIndex = 0;
+//                        foreach($articles as $key => $articleWithCL) {
+//                            if ($articleWithCL['cl'] === ChampLibre::SPECIC_COLLINS_BL) {
+//                                $wantedIndex = $key;
+//                                break;
+//                            }
+//                        }
+//                        $articleArray = $articles[$wantedIndex];
+//                        array_push($response['refs'], $toInsert->getBarCode());
+//                        array_push($response['barcodesLabel'], $this->renderView('article/barcodeLabel.html.twig', [
+//                            'refRef' => $toInsert->getArticleFournisseur()->getReferenceArticle()->getReference(),
+//                            'refLabel' => $toInsert->getArticleFournisseur()->getReferenceArticle()->getLibelle(),
+//                            'artLabel' => $toInsert->getLabel(),
+//                            'artBL' => $wantBL ? $wantBL->getParametre() && $articleArray['cl'] === ChampLibre::SPECIC_COLLINS_BL ? $articleArray['bl'] : null : null,
+//                        ])
+//                        );
+//                        $counter++;
+//                    }
+//                }
+//            }
+//            $em->flush();
+//            return new JsonResponse($response);
+//        }
+//        throw new NotFoundHttpException("404");
+//    }
 
     /**
      * @Route("/apiArticle", name="article_by_reception_api", options={"expose"=true}, methods="GET|POST")
